Move chatConversation tests to the existent test file
